Rename submodule under `cupy.fft` package
