+++ cupy/fft/__init__.py
@@ -1,19 +1,19 @@
-from cupy.fft.fft import fft  # NOQA
-from cupy.fft.fft import fft2  # NOQA
-from cupy.fft.fft import fftfreq  # NOQA
-from cupy.fft.fft import fftn  # NOQA
-from cupy.fft.fft import fftshift  # NOQA
-from cupy.fft.fft import hfft  # NOQA
-from cupy.fft.fft import ifft  # NOQA
-from cupy.fft.fft import ifft2  # NOQA
-from cupy.fft.fft import ifftn  # NOQA
-from cupy.fft.fft import ifftshift  # NOQA
-from cupy.fft.fft import ihfft  # NOQA
-from cupy.fft.fft import irfft  # NOQA
-from cupy.fft.fft import irfft2  # NOQA
-from cupy.fft.fft import irfftn  # NOQA
-from cupy.fft.fft import rfft  # NOQA
-from cupy.fft.fft import rfft2  # NOQA
-from cupy.fft.fft import rfftfreq  # NOQA
-from cupy.fft.fft import rfftn  # NOQA
+from cupy.fft._fft import fft  # NOQA
+from cupy.fft._fft import fft2  # NOQA
+from cupy.fft._fft import fftfreq  # NOQA
+from cupy.fft._fft import fftn  # NOQA
+from cupy.fft._fft import fftshift  # NOQA
+from cupy.fft._fft import hfft  # NOQA
+from cupy.fft._fft import ifft  # NOQA
+from cupy.fft._fft import ifft2  # NOQA
+from cupy.fft._fft import ifftn  # NOQA
+from cupy.fft._fft import ifftshift  # NOQA
+from cupy.fft._fft import ihfft  # NOQA
+from cupy.fft._fft import irfft  # NOQA
+from cupy.fft._fft import irfft2  # NOQA
+from cupy.fft._fft import irfftn  # NOQA
+from cupy.fft._fft import rfft  # NOQA
+from cupy.fft._fft import rfft2  # NOQA
+from cupy.fft._fft import rfftfreq  # NOQA
+from cupy.fft._fft import rfftn  # NOQA
 from cupy.fft import config  # NOQA
