+++ cupyx/scipy/fft/fft.py
@@ -4,10 +4,11 @@ import warnings
 import numpy as np
 
 import cupy
+
 from cupy.cuda import cufft
-from cupy.fft.fft import (_fft, _default_fft_func, hfft as _hfft,
-                          ihfft as _ihfft, _size_last_transform_axis)
-from cupy.fft.fft import fftshift, ifftshift, fftfreq, rfftfreq
+from cupy.fft._fft import (_fft, _default_fft_func, hfft as _hfft,
+                           ihfft as _ihfft, _size_last_transform_axis)
+from cupy.fft import fftshift, ifftshift, fftfreq, rfftfreq
 
 from cupyx.scipy.fftpack import get_fft_plan
